CP-54351 move BaseISCSI from /opt
