+++ tests/test_ISCSISR.py
@@ -1,5 +1,5 @@
 import unittest
-import BaseISCSI
+from sm import BaseISCSI
 import unittest.mock as mock
 from sm.core import xs_errors
 import os
@@ -81,10 +81,10 @@ class NonInitingMultiLUNISCSISR(BaseISCSI.BaseISCSISR):
 @mock.patch('sm.core.xs_errors.XML_DEFS', 'libs/sm/core/XE_SR_ERRORCODES.xml')
 class TestVdiTypeSetting(unittest.TestCase):
 
-    @mock.patch('BaseISCSI.iscsi.discovery', autospec=True)
-    @mock.patch('BaseISCSI.iscsi.ensure_daemon_running_ok', autospec=True)
-    @mock.patch('BaseISCSI.util._testHost', autospec=True)
-    @mock.patch('BaseISCSI.util._convertDNS', autospec=True)
+    @mock.patch('sm.BaseISCSI.iscsi.discovery', autospec=True)
+    @mock.patch('sm.BaseISCSI.iscsi.ensure_daemon_running_ok', autospec=True)
+    @mock.patch('sm.BaseISCSI.util._testHost', autospec=True)
+    @mock.patch('sm.BaseISCSI.util._convertDNS', autospec=True)
     def load_iscsi_sr(self, convertDNS, testHost, ensure_daemon_running_ok,
                       discovery, iscsi_sr):
         iscsi_sr.load(None)
@@ -133,8 +133,8 @@ class TestMultiLUNISCSISR(unittest.TestCase):
         self.assertEqual(node_ip_port, iscsi_sr.tgtidx)
         self.assertEqual(node_ip_port, iscsi_sr.address)
 
-    @mock.patch('BaseISCSI.os.path.exists', autospec=True)
-    @mock.patch('BaseISCSI.iscsi.get_node_records', autospec=True)
+    @mock.patch('sm.BaseISCSI.os.path.exists', autospec=True)
+    @mock.patch('sm.BaseISCSI.iscsi.get_node_records', autospec=True)
     def test_initPaths_actual_path_is_active(
             self,
             mock_get_node_records,
@@ -148,8 +148,8 @@ class TestMultiLUNISCSISR(unittest.TestCase):
 
         self.assertActiveNodeEquals(self.node1, iscsi_sr)
 
-    @mock.patch('BaseISCSI.os.path.exists', autospec=True)
-    @mock.patch('BaseISCSI.iscsi.get_node_records', autospec=True)
+    @mock.patch('sm.BaseISCSI.os.path.exists', autospec=True)
+    @mock.patch('sm.BaseISCSI.iscsi.get_node_records', autospec=True)
     def test_initPaths_active_path_detection(
             self,
             mock_get_node_records,
@@ -173,7 +173,7 @@ class TestMultiLUNISCSISR(unittest.TestCase):
 @mock.patch('sm.core.xs_errors.XML_DEFS', 'libs/sm/core/XE_SR_ERRORCODES.xml')
 class TestISCSISR(unittest.TestCase):
 
-    @mock.patch('BaseISCSI.util._convertDNS', autospec=True)
+    @mock.patch('sm.BaseISCSI.util._convertDNS', autospec=True)
     def test_load_assert_utf_8_chap_credencials(
             self,
             mock__convertDNS):
